:sparkles: feature: user can create short url and end point to create…

+++ src/main/java/com/hackathon/thoven/repositories/GroupInfoJpaRepository.java
@@ -3,7 +3,7 @@ package com.hackathon.thoven.repositories;
 import com.hackathon.thoven.model.GroupInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupInfoJpaRepository extends JpaRepository<GroupInfo, Long> {
+public interface GroupInfoJpaRepository extends JpaRepository<GroupInfo, Integer> {
 
     public GroupInfo getByGroupName(String groupName);
 }
